% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar_close}
\alias{ar_close}
\title{Close Connection to a serial port}
\usage{
ar_close(fd)
}
\arguments{
\item{fd}{File descriptor returned by \code{ar_init()}. Should be an integer.}
}
\description{
This function closes the connection opened by \code{ar_init()}.
}
\examples{
\dontrun{
con <- ar_init("/dev/cu.SLAB_USBtoUART") 
ar_close(con)
}

}
