% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar_flush}
\alias{ar_flush}
\title{Flush serial port}
\usage{
ar_flush(con)
}
\arguments{
\item{con}{File descriptor returned by \code{ar_init()}. Should be an integer.}
}
\description{
clearing the serial port's buffer
}
\examples{
\dontrun{
con <- ar_init("/dev/cu.SLAB_USBtoUART")
ar_flush(con)
}

}
