% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_flush_hard.R
\name{ar_flush_hard}
\alias{ar_flush_hard}
\title{Flush serial port in the hard way}
\usage{
ar_flush_hard(fd, flush_time = 0.05, include_c_flush = TRUE)
}
\arguments{
\item{flush_time}{Time to flush buffed results in the serial channel}

\item{include_c_flush}{T/F for whether to call ar_flush at the beginning}
}
\description{
In theory \code{ar_flush} should work but it didn't work out for me
at least. So I recreated this \code{flush} feature in R, which basically let the
data flow for a short time (default 50 ms). The selection of this value
depends of many factors including the USB port and the size of data so please
pick this value wisely.
}
