% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar_init}
\alias{ar_init}
\title{Initiate a connection to the serial port}
\usage{
ar_init(serialport, baud = 9600L)
}
\arguments{
\item{serialport}{Name of the serial port (e.g. "/dev/tty.usbserial","COM1")}

\item{baud}{Baud rate (bps) in integer.}
}
\value{
If connection is setup successfully, this function will return the
file descriptor. If not, it will return -1.
}
\description{
This function sets up a connection to the serial port at the
specified port name and baud rate. The port is set in 8-N-1 mode and is
opened in fully raw mode so you can send binary data.
}
\examples{
\dontrun{
# On Windows
con <- ar_init("COM1") 

# On Mac
con <- ar_init("/dev/cu.SLAB_USBtoUART")
}

}
