% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ar_is_open}
\alias{ar_is_open}
\title{Check if a file descriptor is still open}
\usage{
ar_is_open(fd)
}
\arguments{
\item{fd}{File descriptor returned by \code{ar_init()}. Should be an integer.}
}
\description{
This function checks whether a stored file descriptor is still
open.
}
\examples{
\dontrun{
con <- ar_init("/dev/cu.SLAB_USBtoUART") 
ar_is_open(con)
}
}
