% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_plotter.R
\name{ar_plotter}
\alias{ar_plotter}
\title{Arduino Plotter}
\usage{
ar_plotter(fd, names = NULL, sep_fun = ar_sep_comma,
  reduce_freq = TRUE, flush_time = 0.05, eolchar = "\\n",
  buf_max = 256, timeout = 5000)
}
\arguments{
\item{names}{Labels for variables}

\item{sep_fun}{A function that separates the inline string into different
variables. By default, it is \code{ar_sep_comma}, which splits the string by
comma. You can easily write your own function, which can even do additional
calculation.}

\item{reduce_freq}{T/F. It might be the case that plotly streaming API can't
process frequency > 25 Hz (40ms delay time) or it might be the case that
my computer doesn't have enough power. Anyway, I set this option here to
add 40ms delay time to reduce the sampling frequency.}

\item{flush_time}{Time to flush buffed results in the serial channel}

\item{eolchar}{End of line character. Default value is \code{'\\n'}}

\item{buf_max}{Maximum length of one line of entry. Default is 256.}

\item{timeout}{Timeout for reads in millisecs. Default is 5000 ms.}
}
\description{
Arduino Plotter
}
