% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_sep.R
\name{ar_sep_comma}
\alias{ar_sep_comma}
\title{Split string by special symbols}
\usage{
ar_sep_comma(x)
}
\arguments{
\item{x}{Arduino Reading. Usually a string and needs to be chopped and
converted to numbers.}
}
\description{
This function determines how the plotter extracts numbers from
the readings. By default, it trims of the "new line" symbol at the end of
the row and split the string by special symbols (if you have used
\code{tidyr::separate}, they are the same). You can write your own function if you
need special processing.
}
