% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_stream.R
\name{ar_stream}
\alias{ar_stream}
\title{Stream serial port data into R console}
\usage{
ar_stream(fd, eolchar = "\\n", buf_max = 256, timeout = 5000)
}
\arguments{
\item{fd}{File descriptor returned by `ar_init()`. Should be an integer.}

\item{eolchar}{End of line character. Default value is `'\\n'`}

\item{buf_max}{Maximum length of one line of entry. Default is 256.}

\item{timeout}{Timeout for reads in millisecs. Default is 5000 ms.}
}
\description{
This function wraps around `ar_read()` and will read serial
port data into R console until user stop it.
}
