% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar_ws_plotter.R
\name{ar_ws_plotter}
\alias{ar_ws_plotter}
\title{Arduino Plotter for websocket}
\usage{
ar_ws_plotter(ws, names = NULL, sep_fun = ar_sep_comma, freq = 50)
}
\arguments{
\item{ws}{websocket connection}

\item{names}{Labels for variables}

\item{sep_fun}{A function that separates the inline string into different
variables. By default, it is \code{ar_sep_comma}, which splits the string by
comma. You can easily write your own function, which can even do additional
calculation.}

\item{freq}{Length of miliseconds that the shiny app will check for new data.
The default 50ms (20Hz) is almost the maxium on normal computer nowadays.}
}
\description{
Arduino Plotter for websocket
}
